Fix failing tests
Problem:
Unit tests were failing due to the base frame is not properly printed. In case
a base frame is created by Windowtester the frame is wrapped in a window. This
renders the frame and all components on it invisible.

Solution:
Use the base frame directly and do not wrap it in a window. Any Frame/JFrame is
already an instance of Window class.