Fix failing test
Problem:
JDK 21 marks the Security Manager mechanism of the JDK as deprecated and about to be removed. Setting the SM throws an UnsupportedOperationException.

Solution:
Bypass the setting of an SM by default the system property to false which renders the guard also false skipping the SM initialization during test execution.
User of the library can still set the use the SM mechanism by setting the system property to "true".