Revert changes
Problem:
The deployment to Maven Central fails. Re-reading the documentation on how to
publish the previous approach seems to be correct.
Try to use an authentication consisting of username and password generated on
sonatype.

Solution:
Regenerate user token which consists of a generated username and a password.
This complies with the description given in the sonatype documentation:
https://central.sonatype.org/publish/publish-portal-maven/#credentials